logger/requirements.txt
opcua
psycopg2-binary
pyyaml
cryptography
pandas
sqlalchemy